Server : Updated the schemas, added hooks to handle deletion and remov…

+++ server/schema/types/lyric.js
@@ -1,10 +1,12 @@
+import { GraphQLError } from "graphql";
 import { Lyric, Song } from "../../models/index.js";
 
 export const typeDef = `#graphql
   type Lyric {
-    id: String!
+    id: ID!
     song: Song
     likes: Int
+    isLiked: Boolean
     content: String!
   }
 
@@ -45,7 +47,14 @@ export const resolver = {
     lyric: async (_, { id }) => await Lyric.findById(id),
   },
   Mutation: {
-    addLyric: async (_, { content, songId }) => {
+    addLyric: async (_, { content, songId }, { user }) => {
+
+      if (!user) throw new GraphQLError("Not Authenticated", {
+        extensions: { code: 'UNAUTHENTICATED', http: {
+          status: 401,
+        }},
+      });
+
       const [lyric] = await Song.addLyric(songId, content);
 
       return {
@@ -55,8 +64,14 @@ export const resolver = {
         lyric,
       }
     },
-    deleteLyric: async (_, { id }) => {
+    deleteLyric: async (_, { id }, { user }) => {
       try {
+        if (!user) throw new GraphQLError("Not Authenticated", {
+          extensions: { code: 'UNAUTHENTICATED', http: {
+            status: 401,
+          }},
+        });
+
         const lyric = await Lyric.findOneAndDelete({ _id: id });
         return {
           code: "200",
@@ -73,8 +88,15 @@ export const resolver = {
         }
       }
     },
-    likeLyric: async (_, { id }) => {
-      const res = await Lyric.like(id);
+    likeLyric: async (_, { id }, { user }) => {
+
+      if (!user) throw new GraphQLError("Not Authenticated", {
+        extensions: { code: 'UNAUTHENTICATED', http: {
+          status: 401,
+        }},
+      });
+
+      const res = await Lyric.likeDislike(user.id, id);
       console.log(res);
       return {
         code: "200",
@@ -85,6 +107,12 @@ export const resolver = {
     }
   },
   Lyric: {
+    likes: (parentValue) => {
+      return parentValue.likedBy.length;
+    },
+    isLiked: (parentValue, _, { user }) => {
+      return parentValue.likedBy.includes(user?.id);
+    },
     song: async (parentValue) => {
       const lyric = await Lyric.findById(parentValue.id).populate('song');
       return lyric.song;
